Setup `@ContributedBinding` Anvil example
